+++ stencil-workspace/src/components/modus-date-picker/utils/modus-date-picker.types.ts
@@ -1,0 +1,10 @@
+import { ModusDateInputCustomEvent } from '../../../components';
+
+export type DateInputType = 'start' | 'end' | 'single';
+
+export type DateInputEventData = {
+  date: Date;
+  type: DateInputType;
+};
+
+export type DateInputEvent = ModusDateInputCustomEvent<DateInputEventData>;